requirements.txt
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
